Initialize toggle state on launch
This is a little hacky, but it is needed so that the interface does not
start in a weird state where all components are visible. E.g., when
sorting, we want to either sort by attribute or by expression, and in
one of these modes, we don't want inputs for the other modes be visible.