Name purpose UI (#237)
* Move name and description out of save area

* Fix loading from savefile loop in definition view

* Put purpose statement after type contract

* Remove dropdown for save button

The textboxes that were previously in the save section were moved
elsewhere. We no longer need a dropdown for a single button.

* Remove purpose statement from top of saved view

* Add different placeholders for purpose statements

* Remove unused import

* Make purpose statement configurable

* Support two kinds of sort (#238)

* Add toggle component which toggle other components

This helps with creating a dropdown for two kinds of sorting behaviors
within one Sort transformer in order to support sorting by attribute and
sorting by expression. These behaviors require different inputs, and we
want to hide the inputs we are not currently using without switching
transformers.

* Support two kinds of sorting

Sorting by expression and sorting by attribute

* Add type inference for columns

The main purpose of this is to correctly interpret numeric strings or
boolean strings.

* Use type inference when sorting by attribute

* Fix bug in sort toggle

Make sorting by expression actually hide the attribute dropdown

* Initialize toggle state on launch

This is a little hacky, but it is needed so that the interface does not
start in a weird state where all components are visible. E.g., when
sorting, we want to either sort by attribute or by expression, and in
one of these modes, we don't want inputs for the other modes be visible.

* Remove redundant option for default toggle select

Instead of using the `Select` component, use a custom select with no
default value. This is because a valid toggle option must be chosen at
all times, so there is no need for a separate default option.

* Reformat

* Update inline documentation for Sort to reflect two variants

* Use consistent language for formula titles (#241)

* Use consistent language for formula titles

* Update Build + Transform's language for formulas as well

* Remove `description` field from merge

Co-authored-by: Thomas Castleman <thomascastleman0@gmail.com>

* Increase default plugin height, decrease purpose statement height

* Initial draft of custom transformer name placeholders

* Name component does not need a title

* Update placeholders to be based on Animals dataset

* Prevent saving transformers with duplicate names (#245)

* Add duplicate prevention functionality

* Remove extra console log

* Add name dupe protection for saved transformers

* Trim whitespace from saved transformer names

* Fix whitespace not being trimmed from textbox

Co-authored-by: Thomas Castleman <thomascastleman0@gmail.com>
Co-authored-by: Paul Biberstein <40366824+P-bibs@users.noreply.github.com>